Use model definition of TestTreeSettings in Experimental GUI
