+++ src/Experimental/experimental-gui/Presenters/TreeDisplayStyle.cs
@@ -1,4 +1,4 @@
-﻿// ***********************************************************************
+// ***********************************************************************
 // Copyright (c) 2016 Charlie Poole
 //
 // Permission is hereby granted, free of charge, to any person obtaining
@@ -21,7 +21,7 @@
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // ***********************************************************************
 
-namespace TestCentric.Gui.Settings
+namespace TestCentric.Gui.Presenters
 {
     /// <summary>
     /// Indicates how a tree should be displayed
