+++ src/Experimental/experimental-gui/Views/SettingsPages/GuiSettingsPage.cs
@@ -1,4 +1,4 @@
-﻿// ***********************************************************************
+// ***********************************************************************
 // Copyright (c) 2015 Charlie Poole
 //
 // Permission is hereby granted, free of charge, to any person obtaining
@@ -41,7 +41,7 @@ namespace TestCentric.Gui.Views.SettingsPages
 
         public override void ApplySettings()
         {
-            Settings.Gui.TestTree.InitialTreeDisplay = (TreeDisplayStyle)initialDisplayComboBox.SelectedIndex;
+            Settings.Gui.TestTree.InitialTreeDisplay = initialDisplayComboBox.SelectedIndex;
             Settings.Gui.TestTree.SaveVisualState = saveVisualStateCheckBox.Checked;
         }
